Slater determinants

Slater determinants

Slater determinants are wavefunctions constructed from anti-symmetrized one-particle states.

Construction of Slater determinants

EnergyExpressions.SlaterDeterminant — Type.

SlaterDeterminant(orbitals::Vector{O})

Constructs a Slater determinant from a set of spin-orbitals.

Examples

julia> SlaterDeterminant([:a, :b])
a(1)b(2) - a(2)b(1)

julia> SlaterDeterminant([:a, :b, :c])
a(1)b(2)c(3) - a(1)b(3)c(2) - a(2)b(1)c(3) + a(2)b(3)c(1) + a(3)b(1)c(2) - a(3)b(2)c(1)

source

EnergyExpressions.SlaterDeterminant — Method.

SlaterDeterminant(configuration::Configuration{<:SpinOrbital})

Constructs a Slater determinant from the spin-orbitals of the spin-configuration configuration.

Examples

julia> SlaterDeterminant(spin_configurations(c"1s2")[1])
1s₀α(1)1s₀β(2) - 1s₀α(2)1s₀β(1)

source

Base.length — Method.

length(slater_determinant)

Return the number of spin-orbitals in the Slater determinant.

source

EnergyExpressions.AdjointSlaterDeterminant — Type.

AdjointSlaterDeterminant(slater_determinant)

Representation of the Hermitian conjugate (dual vector) of a Slater determinant. Constructed via the usual adjoint operator.

source

Base.adjoint — Method.

adjoint(slater_determinant)

Construct the adjoint of slater_determinant

Examples

julia> SlaterDeterminant([:a, :b])'
[a(1)b(2) - a(2)b(1)]†

source

Base.length — Method.

length(adjoint_slater_determinant)

Return the number of spin-orbitals in the adjoint Slater determinant.

source

N-body operators

EnergyExpressions.NBodyOperator — Type.

NBodyOperator{N}

N-body operator coupling N bodies each between two Slater determinants

Examples

julia> ZeroBodyOperator
Ω₀

julia> OneBodyOperator
Ω₁

julia> TwoBodyOperator
Ω₂

julia> NBodyOperator{6}
Ω₆

source

EnergyExpressions.NBodyTermFactor — Type.

NBodyTermFactor

Abstract type for a factor in a term in a N-body matrix element expansion

source

EnergyExpressions.OrbitalOverlap — Type.

OrbitalOverlap(a,b)

Represents the overlap between the orbitals a and b in a N-body matrix element expansion.

Examples

julia> EnergyExpressions.OrbitalOverlap(:a,:b)
⟨a|b⟩

source

EnergyExpressions.OrbitalMatrixElement — Type.

OrbitalMatrixElement(a,o,b)

Represents the N-body matrix element between the sets of orbitals a and b.

Examples

julia> EnergyExpressions.OrbitalMatrixElement((:a,:b), TwoBodyOperator, (:c,:d))
⟨a b|Ω₂|c d⟩

source

Example usage

julia> sa = SlaterDeterminant([:l, :a])
l(1)a(2) - l(2)a(1)

julia> sb = SlaterDeterminant([:k, :b])
k(1)b(2) - k(2)b(1)

julia> sa'ZeroBodyOperator*sb
+ ⟨l|k⟩⟨a|b⟩ - ⟨l|b⟩⟨a|k⟩ - ⟨a|k⟩⟨l|b⟩ + ⟨a|b⟩⟨l|k⟩

julia> sa'OneBodyOperator*sb
+ ⟨l|Ω₁|k⟩⟨a|b⟩ + ⟨l|k⟩⟨a|Ω₁|b⟩ - ⟨l|Ω₁|b⟩⟨a|k⟩ - ⟨l|b⟩⟨a|Ω₁|k⟩ - ⟨a|Ω₁|k⟩⟨l|b⟩ - ⟨a|k⟩⟨l|Ω₁|b⟩ + ⟨a|Ω₁|b⟩⟨l|k⟩ + ⟨a|b⟩⟨l|Ω₁|k⟩